requirements.txt
Django==2.1.4
djangorestframework==3.9.0
freezegun==0.3.11
